move info command function
